Activate stdin handler processing
The `close` handler when `stdin` is `end`ed is not being called.  Calling `resume()` on `process.stdin` will enable that handler so it will be called properly.  It should 'probably' be called earlier in the program, perhaps at the very start, but if it spools up fast enough it should not matter, and this is the same place similar code in other projects places it as well.